@@ -1,6 +1,6 @@
 module github.com/home-assistant/os-agent
 
-go 1.15
+go 1.19
 
 require (
 	github.com/coreos/go-systemd/v22 v22.5.0
@@ -10,4 +10,9 @@ require (
 	github.com/godbus/dbus/v5 v5.1.0
 	github.com/natefinch/atomic v1.0.1
 	github.com/opencontainers/runtime-spec v1.1.0
+)
+
+require (
+	golang.org/x/sys v0.6.0 // indirect
+	golang.org/x/text v0.8.0 // indirect
 )
